Flatten steps organization
